New temp dir parameter and print of parameters during test

+++ test/src/test/java/org/corfudb/AbstractCorfuTest.java
@@ -32,7 +32,6 @@ import static org.fusesource.jansi.Ansi.ansi;
 public class AbstractCorfuTest {
 
     public Set<Callable<Object>> scheduledThreads;
-    public Set<String> temporaryDirectories;
     public String testStatus = "";
 
     public static final CorfuTestParameters PARAMETERS =
@@ -65,18 +64,26 @@ public class AbstractCorfuTest {
         }
     };
 
-    public static void deleteFolder(File folder) {
+    /** Delete a folder.
+     *
+     * @param folder        The folder, as a File.
+     * @param deleteSelf    True to delete the folder itself,
+     *                      False to delete just the folder contents.
+     */
+    public static void deleteFolder(File folder, boolean deleteSelf) {
         File[] files = folder.listFiles();
         if (files != null) { //some JVMs return null for empty dirs
             for (File f : files) {
                 if (f.isDirectory()) {
-                    deleteFolder(f);
+                    deleteFolder(f, true);
                 } else {
                     f.delete();
                 }
             }
         }
-        folder.delete();
+        if (deleteSelf) {
+            folder.delete();
+        }
     }
 
     @Before
@@ -89,10 +96,6 @@ public class AbstractCorfuTest {
         scheduledThreads = new HashSet<>();
     }
 
-    @Before
-    public void setupTempDirs() {
-        temporaryDirectories = new HashSet<>();
-    }
 
     @After
     public void cleanupScheduledThreads() {
@@ -102,19 +105,14 @@ public class AbstractCorfuTest {
         scheduledThreads.clear();
     }
 
-    public String getTempDir() {
-        String tempdir = com.google.common.io.Files.createTempDir().getAbsolutePath();
-        temporaryDirectories.add(tempdir);
-        return tempdir;
+    /** Clean the per test temporary directory (PARAMETERS.TEST_TEMP_DIR)
+     */
+    @After
+    public void cleanPerTestTempDir() {
+        deleteFolder(new File(PARAMETERS.TEST_TEMP_DIR),
+                false);
     }
 
-    @After
-    public void deleteTempDirs() {
-        for (String s : temporaryDirectories) {
-            File folder = new File(s);
-            deleteFolder(folder);
-        }
-    }
 
     public void calculateAbortRate(int aborts, int transactions) {
         final float FRACTION_TO_PERCENT = 100.0F;
